add new fonctions of ListMovie in reaction with user

+++ src/Controller/ListMovieController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Movie;
 use App\Entity\ListMovie;
+use App\Entity\User;
 use App\Form\ListMovieType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -14,7 +15,9 @@ use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\View\View;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityNotFoundException;
+use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Doctrine\Common\Collections\ArrayCollection;
 
 
@@ -33,15 +36,20 @@ final class ListMovieController extends FOSRestController
 		//create form
 		$form = $this->createForm(ListMovieType::class, $listMovie);//build form
 		//handle request
-		$data = json_decode($request->getContent(), true);// true for get table relative	
-		
+		$data = json_decode($request->getContent(), true);// true for get table relative			
+        
 		//submit data
 		$form->submit($data);//remplir table
 		//if (not valid)
 		if(!$form->isValid()) {
 			throw new BadRequestHttpException($form->getErrors(true));
-		}	
-			
+		}
+        
+        //find current user & setUser
+        $user = $this->getUser();
+        //print_r($user);
+        $listMovie->setUser($user);
+        
 		//saving entity && return 
 		$em->persist($listMovie);
 		$em->flush();
@@ -73,19 +81,29 @@ final class ListMovieController extends FOSRestController
 		if (!$listMovie) {
             throw new EntityNotFoundException('ListMovie with id '.$listMovieId.' does not exist !');
         }
-		
-		$form = $this->createForm(ListMovieType::class, $listMovie);//build form
-		$data = json_decode($request->getContent(), true);// true for get table relative	
-		$form->submit($data);//remplir table
-		if(!($form->isValid()&&$form->isSubmitted())) {
-			throw new BadRequestHttpException($form->getErrors(true));
-		}
-		
-		$em->persist($listMovie);
-		$em->flush();
-		
-		// In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
-		return View::create($listMovie, Response::HTTP_OK);
+        //find current user 
+        $user = $this->getUser();
+        //find user of list
+        $listUser = $listMovie->getUser(); 
+        
+        if ($user == $listUser) {          
+                    
+            $form = $this->createForm(ListMovieType::class, $listMovie);//build form
+            $data = json_decode($request->getContent(), true);// true for get table relative	
+            $form->submit($data);//remplir table
+            if(!($form->isValid()&&$form->isSubmitted())) {
+                throw new BadRequestHttpException($form->getErrors(true));
+            }
+            
+            $em->persist($listMovie);
+            $em->flush();
+            
+            // In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
+            return View::create($listMovie, Response::HTTP_OK);
+        }
+        else {
+            throw new AccessDeniedException("Access Denied");
+        }
 	}
 	
 	/**
@@ -98,17 +116,28 @@ final class ListMovieController extends FOSRestController
 		if (!$listMovie) {
             throw new EntityNotFoundException('ListMovie with id '.$listMovieId.' does not exist!!!');
         }
+        
+		$user = $this->getUser();
+        $listUser = $listMovie->getUser(); 
+		if ($user == $listUser) {
+            $form = $this->createForm(ListMovieType::class, $listMovie);//build form
+            $data = json_decode($request->getContent(), true);// true for get table relative
+            $form->submit($data,false);//remplir table
+            if(!$form->isValid()) {
+                throw new BadRequestHttpException($form->getErrors(true));
+            }
+            
+            $em->persist($listMovie);
+            $em->flush();
 		
-		$form = $this->createForm(ListMovieType::class, $listMovie);//build form
-		$data = json_decode($request->getContent(), true);// true for get table relative
-		$form->submit($data,false);//remplir table
-		if(!$form->isValid()) {
-			throw new BadRequestHttpException($form->getErrors(true));
-		}
-		$em->persist($listMovie);
-		$em->flush();
-		//In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
-		return View::create($listMovie, Response::HTTP_OK);
+            //In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
+            return View::create($listMovie, Response::HTTP_OK);
+
+        }
+        else {
+            throw new AccessDeniedException("Access Denied");
+        }		
+
 	}
 	
 	/**
@@ -121,12 +150,20 @@ final class ListMovieController extends FOSRestController
             throw new EntityNotFoundException('ListMovie with id '.$listMovieId.' does not exist!!!');
         }
 		
-		$em->remove($listMovie);
-		$em->flush();
-		
-		// In case our DELETE was a success we need to return a 204 HTTP NO CONTENT response. The object is deleted.
-		return View::create(["delete"], Response::HTTP_NO_CONTENT);
-		
+        $user = $this->getUser();
+        $listUser = $listMovie->getUser(); 
+		if ($user == $listUser) {
+            
+            $em->remove($listMovie);
+            $em->flush();
+            
+            // In case our DELETE was a success we need to return a 204 HTTP NO CONTENT response. The object is deleted.
+            return View::create(["delete"], Response::HTTP_NO_CONTENT);
+		}
+        else {
+            throw new AccessDeniedException("Access Denied");
+        }
+        
 	}
 	
 	/**
@@ -142,21 +179,28 @@ final class ListMovieController extends FOSRestController
             throw new EntityNotFoundException('ListMovie with id '.$listMovieId.' does not exist!');
         }
 		
-		$movieIds = $request->request->all();//aussi getContent()
-		foreach ($movieIds as $movieId) {
-			$movieTemp = $em->getRepository(Movie::class)->find($movieId);
-				if (!$movieTemp) {
-					throw new EntityNotFoundException('Movie with id '.$movieId.' does not exist!!!');
-				}
-			$listMovie->addMovie($movieTemp);
+        $user = $this->getUser();
+        $listUser = $listMovie->getUser(); 
+		if ($user == $listUser) {            
+            $movieIds = $request->request->all();//aussi getContent()
+            foreach ($movieIds as $movieId) {
+                $movieTemp = $em->getRepository(Movie::class)->find($movieId);
+                    if (!$movieTemp) {
+                        throw new EntityNotFoundException('Movie with id '.$movieId.' does not exist!!!');
+                    }
+                $listMovie->addMovie($movieTemp);
+            }
+            
+            $em->persist($listMovie);
+            $em->flush();
+            
+            //In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
+            return View::create($listMovie, Response::HTTP_OK);
 		}
-		
-		$em->persist($listMovie);
-		$em->flush();
-		
-		//In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
-		return View::create($listMovie, Response::HTTP_OK);
-		
+        else {
+            throw new AccessDeniedException("Access Denied");
+        }
+        
     }
 	
 	/**
@@ -170,21 +214,105 @@ final class ListMovieController extends FOSRestController
 		$listMovie = $em->getRepository(ListMovie::class)->find($listMovieId);		
 		if (!$listMovie) {
             throw new EntityNotFoundException('ListMovie with id '.$listMovieId.' does not exist!');
-        }		
-
-		$listOfFilm = $listMovie->getMovies();
-		$movieIds = $request->request->all();
-		foreach ($listOfFilm as $filmTemp) {
-			if (in_array($filmTemp->getId(),$movieIds)) {
-				$listMovie->removeMovie($filmTemp);
-			}
-		}
+        }
 		
-		$em->persist($listMovie);
-		$em->flush();
-		
-		//In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
-		return View::create($listMovie, Response::HTTP_OK);	
+        $user = $this->getUser();
+        $listUser = $listMovie->getUser(); 
+		if ($user == $listUser) { 
+            $listOfFilm = $listMovie->getMovies();
+            $movieIds = $request->request->all();
+            foreach ($listOfFilm as $filmTemp) {
+                if (in_array($filmTemp->getId(),$movieIds)) {
+                    $listMovie->removeMovie($filmTemp);
+                }
+            }
+            
+            $em->persist($listMovie);
+            $em->flush();
+            
+            //In case our PUT was a success we need to return a 200 HTTP OK response with the object as a result of PUT
+            return View::create($listMovie, Response::HTTP_OK);	
+        }
+        else {
+            throw new AccessDeniedException("Access Denied");
+        }
     }
 	
+    /**
+	 * find in AllListMovie
+	 */
+	 public function findAllListMovie(Request $request, EntityManagerInterface $em): View
+	{		
+		$search = [];
+		
+		if($request->query->get('name')) {
+			$search['name'] = $request->query->get('name');
+		}
+		if($request->query->get('user')) {
+            $userId = $request->query->get('user');
+			$search['user'] =  $em->getRepository(User::class)->find($userId);
+		}
+		$listMovies = $em->getRepository(ListMovie::class)->findBy($search);
+
+		// In case our GET was a success we need to return a 200 HTTP OK response with the request object
+		return View::create($listMovies, Response::HTTP_OK);	
+	}
+
+    /**
+	 * findLike in AllListMovie
+	 */
+	 public function findLikeAllListMovie(Request $request, EntityManagerInterface $em): View
+	{		
+		$search = [];
+
+        if($request->query->get('keyWord')) {
+			$search['keyWord'] = $request->query->get('keyWord');
+            
+            if($request->query->get('user')) {
+                $search['user'] = $request->query->get('user');
+                $listMovies = $em->getRepository(ListMovie::class)->findLikeId($search['keyWord'],$search['user']);
+            }
+            else {
+			$listMovies = $em->getRepository(ListMovie::class)->findLikeAll($search['keyWord']);
+            }
+        }
+		// In case our GET was a success we need to return a 200 HTTP OK response with the request object
+		return View::create($listMovies, Response::HTTP_OK);	
+	}
+
+
+    /**
+	 * find in MyListMovie
+	 */
+	 public function findMyListMovie(Request $request, EntityManagerInterface $em): View
+	{		
+		$search = [];
+		
+        $search['user'] = $this->getUser();
+		if($request->query->get('name')) {
+			$search['name'] = $request->query->get('name');
+		}
+		$listMovies = $em->getRepository(ListMovie::class)->findBy($search);
+		
+		// In case our GET was a success we need to return a 200 HTTP OK response with the request object
+		return View::create($listMovies, Response::HTTP_OK);	
+	}
+    
+    /**
+	 * findLike in MyListMovie
+	 */
+	 public function findLikeMyListMovie(Request $request, EntityManagerInterface $em): View
+	{		
+		$search = [];
+		
+        $search['userId'] = $this->getUser()->getId();
+        if($request->query->get('keyWord')) {
+			$search['keyWord'] = $request->query->get('keyWord');
+			$listMovies = $em->getRepository(ListMovie::class)->findLikeId($search['keyWord'],$search['userId']);
+		}
+       
+		// In case our GET was a success we need to return a 200 HTTP OK response with the request object
+		return View::create($listMovies, Response::HTTP_OK);	
+	}
+    
 }
